pyproject.toml
[tool.poetry]
name = "pdftopptx"
version = "0.1.0"
description = ""
authors = ["naoya-matsuno <naoya.matsuno.r1@dc.tohoku.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pypdf2 = "^2.11.1"
python-pptx = "^0.6.21"
pdf2image = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
